Use the shared PadMask for the wakeup pad sets
